Переход к компонентному подходу ООП
